fix(api): consolidate license verify+issue into [action].ts (Hobby plan limit)
Vercel Hobby plan caps Serverless Functions per deployment at 12. PR #39
pushed us to 13 (catalog, checkout, favicon, feed, feedback, health, icon,
license/issue, license/verify, page, stats-sync, stripe/webhook, sync),
which surfaced as a silent "exceeded_serverless_functions_per_deployment"
error after the build step. Source:
  https://vercel.com/docs/limits (Functions Created per Deployment: ∞ on
  Pro/Enterprise; framework-dependent on Hobby — 12 for Node.js without
  Next.js, per the API error message).

Fix: collapse api/license/verify.ts + api/license/issue.ts into a single
Vercel dynamic-route file api/license/[action].ts. The wrapper inspects
URL.pathname and dispatches to the same shared handlers we already had.
Same behavior, same tests, same URLs (/api/license/verify and
/api/license/issue both resolve to the dispatcher) — only the file
boundary changed.

Function count: 13 → 12 (at the ceiling, room for one more before we
either have to upgrade to Pro or consolidate again).

Tests:
- tests/server.test.ts imports both wrappers' SUPPORTED_METHODS contract
  tests against the same shared module. Aliased to keep test names
  meaningful (vercelLicenseVerifyExports / vercelLicenseIssueExports).
- Wrapper structural assertions updated to read [action].ts and verify
  both 'verify' and 'issue' action strings appear in the source — a
  regression dropping one branch is caught even if the other branch's
  test traffic happens to pass.
- npm test → 1753 pass
- npx tsc --noEmit → clean
